% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_number.R
\name{as_number}
\alias{as_number}
\title{Takes a string representing a numerical value, and removes all comma delimiters and \% signs, before converting to a numerical value (an integer or decimal).}
\usage{
as_number(x)
}
\arguments{
\item{x}{A formatted number string}
}
\description{
Takes a string representing a numerical value, and removes all comma delimiters and \% signs, before converting to a numerical value (an integer or decimal).
}
\examples{
as_number("1,000")  #  returns 1000
as_number("10\%")  #  returns 10
}
